feat: @name works from the chat UI, resolved by the daemon
The REPL's @name sugar was client-side: it resolved against Pi's command
list and rewrote the input into plain prompt text before sending. The
chat UI sends prompts through the daemon, which forwarded them verbatim,
so @exa or @web-research typed there reached the model as literal text.

The resolution now lives in core::plugin, and the daemon applies it to
any prompt starting with @, so both frontends mean the same thing by it.
A bare @plugin answers with a description over a new `mention` SSE event
instead of spending a model turn, and an unresolved name answers the
same way with what is available. A plugin command is acked rather than
answered, so the read loop learns to stop on that ack instead of waiting
for an agent_settled that never comes.

GET /v1/tilekit/agent/commands lists everything @name can reach, with
kind and description, which is what a mention picker needs to offer
plugins the way it offers files.
